[refactor]S3 package 생성 #39
